+++ src/headless/index.js
@@ -37,6 +37,7 @@ export { MUCMessage, MUCMessages, MUC, MUCOccupant, MUCOccupants } from './plugi
 
 import './plugins/ping/index.js'; // XEP-0199 XMPP Ping
 import './plugins/pubsub/index.js'; // XEP-0060 Pubsub
+import './plugins/time/index.js'; // XEP-0202 Entity Time
 export {
     MicroblogProfile,
     PubSubFeed,