+++ src/headless/types/plugins/time/types.d.ts
@@ -1,0 +1,5 @@
+export type EntityTime = {
+    utc: Date;
+    tzo: string;
+};
+//# sourceMappingURL=types.d.ts.map