+++ src/index.js
@@ -48,6 +48,7 @@ import './plugins/roomslist/index.js'; // Show currently open chat rooms
 import './plugins/rootview/index.js';
 import './plugins/rosterview/index.js';
 import './plugins/singleton/index.js';
+import './plugins/time-views/index.js'; // Views for XEP-0202 Entity Time
 import './plugins/dragresize/index.js'; // Allows chat boxes to be resized by dragging them
 import './plugins/fullscreen/index.js';
 import './plugins/reactions-views/index.js'; // XEP-0444 Reactions