+++ src/types/plugins/time-views/templates/entity-time-alert.d.ts
@@ -1,0 +1,3 @@
+declare function _default(el: import("../entity-time-alert").default): import("lit-html").TemplateResult<1>;
+export default _default;
+//# sourceMappingURL=entity-time-alert.d.ts.map